requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-google-genai==1.0.10
langchain-qdrant==0.1.4
qdrant-client==1.11.1
tiktoken==0.7.0
pypdf==4.3.1
unstructured
rank-bm25==0.2.2
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
loguru==0.7.2
deepeval==1.1.6
fastapi==0.115.0
uvicorn==0.30.6
tenacity==8.3.0
sentence-transformers==3.0.1
torch
tabulate==0.9.0
streamlit==1.38.0
python-multipart==0.0.9
httpx==0.27.2
langsmith==0.1.121